refactor: migrate from MongoDB to TimescaleDB for candle data storage

+++ src/main/java/com/dzenthai/cryptora/model/dto/Average.java
@@ -3,22 +3,20 @@ package com.dzenthai.cryptora.model.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 
-import java.math.BigDecimal;
-
 
 @Builder
 public record Average(
         @JsonProperty("open_price")
-        BigDecimal openPrice,
+        double openPrice,
         @JsonProperty("close_price")
-        BigDecimal closePrice,
+        double closePrice,
         @JsonProperty("high_price")
-        BigDecimal highPrice,
+        double highPrice,
         @JsonProperty("low_price")
-        BigDecimal lowPrice,
+        double lowPrice,
         @JsonProperty("trade_price")
-        BigDecimal tradePrice,
+        double tradePrice,
         @JsonProperty("price_range")
-        BigDecimal priceRange
+        double priceRange
 )
 {}
