+++ src/main/java/com/dzenthai/cryptora/model/dto/Current.java
@@ -3,29 +3,28 @@ package com.dzenthai.cryptora.model.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 
 
 @Builder
 public record Current(
         @JsonProperty("open_price")
-        BigDecimal openPrice,
+        double openPrice,
         @JsonProperty("close_price")
-        BigDecimal closePrice,
+        double closePrice,
         @JsonProperty("high_price")
-        BigDecimal highPrice,
+        double highPrice,
         @JsonProperty("low_price")
-        BigDecimal lowPrice,
+        double lowPrice,
         @JsonProperty("volume")
-        BigDecimal volume,
+        double volume,
         @JsonProperty("amount")
-        BigDecimal amount,
+        double amount,
         @JsonProperty("open_time")
         Instant openTime,
         @JsonProperty("close_time")
         Instant closeTime,
         @JsonProperty("trades")
-        Long trades
+        long trades
 ) {
 }
