+++ src/main/java/com/dzenthai/cryptora/model/dto/Min.java
@@ -3,24 +3,22 @@ package com.dzenthai.cryptora.model.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 
-import java.math.BigDecimal;
-
 
 @Builder
 public record Min(
         @JsonProperty("open_price")
-        BigDecimal openPrice,
+        double openPrice,
         @JsonProperty("close_price")
-        BigDecimal closePrice,
+        double closePrice,
         @JsonProperty("high_price")
-        BigDecimal highPrice,
+        double highPrice,
         @JsonProperty("low_price")
-        BigDecimal lowPrice,
+        double lowPrice,
         @JsonProperty("price_range")
-        BigDecimal priceRange,
+        double priceRange,
         @JsonProperty("volume")
-        BigDecimal volume,
+        double volume,
         @JsonProperty("amount")
-        BigDecimal amount
+        double amount
 ) {
 }
