+++ src/main/java/com/dzenthai/cryptora/model/dto/Total.java
@@ -2,13 +2,12 @@ package com.dzenthai.cryptora.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
-import java.math.BigDecimal;
 
 @Builder
 public record Total(
         @JsonProperty("volume")
-        BigDecimal volume,
+        double volume,
         @JsonProperty("amount")
-        BigDecimal amount
+        double amount
 )
 {}
